+++ HISTORY.rst
@@ -3,7 +3,15 @@
 History
 -------
 
-0.1.0 (2015-07-10)
+0.2.0 (unreleased)
 ++++++++++++++++++
 
-* First release on PyPI.
+* Added high latency mode support.
+  Default value is ``false`` unless ``WLCMANAGER_HIGHLATENCY_DEFAULT`` is set
+  in ``settings``. You can set it before running migrations so existing APs
+  will be updated correctly.
+
+0.1.0 (2015-07-27)
+++++++++++++++++++
+
+* First release.